+++ web/src/main/java/com/nicico/sales/web/controller/InvoiceInternalFormController.java
@@ -39,9 +39,9 @@ public class InvoiceInternalFormController {
         Map<String, Object> params = new HashMap<>();
         params.put(ConstantVARs.REPORT_TYPE, type);
         params.put("ID", rowId);
-        Float mablaghKol = invoiceInternalService.get(Long.valueOf(rowId)).getMablaghKol();
-        Float payForAvarezMalyat = invoiceInternalService.get(Long.valueOf(rowId)).getPayForAvarezMalyat();
-        Float sum = mablaghKol + payForAvarezMalyat;
+        Double mablaghKol = invoiceInternalService.get(Long.valueOf(rowId)).getMablaghKol();
+        Double payForAvarezMalyat = invoiceInternalService.get(Long.valueOf(rowId)).getPayForAvarezMalyat();
+        Double sum = mablaghKol + payForAvarezMalyat;
         params.put("sum_number_to_string", dateUtil.numberToString(String.format ("%.0f", sum)));
         reportUtil.export("/reports/invoice_dakheli.jasper", params, response);
     }
